Apartados 2 y 3 pero el boton Periodic no funciona

+++ HTTP_Server/adc.h
@@ -2,5 +2,5 @@
 #ifndef __ADC_H
 	void ADC1_pins_F429ZI_config(void);
 	int ADC_Init_Single_Conversion(ADC_HandleTypeDef *, ADC_TypeDef  *);
-	float ADC_getVoltage(ADC_HandleTypeDef * , uint32_t );
+	uint32_t ADC_getVoltage(ADC_HandleTypeDef * , uint32_t );
 #endif
